pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Second Project"
authors = ["lovedr6s <praense@icloud.com>"]
readme = "README.md"
packages = [
  { include = "gendiff_code" },
]

[tool.poetry.dependencies]
python = ">=3.8.1"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
gendiff = "gendiff_code.gendiff:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
pyyaml = "^6.0.2"